+++ src/main/java/fr/istic/tp3spring/rest/ProfessionalController.java
@@ -5,11 +5,13 @@ import fr.istic.tp3spring.dao.ProfessionalDAO;
 import fr.istic.tp3spring.domain.Professional;
 import fr.istic.tp3spring.dto.ProfessionalDTO;
 import fr.istic.tp3spring.dto.mapper.MapStructMapper;
-import org.springframework.stereotype.Controller;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Logger;
 
 @RestController
 @RequestMapping("professional")
@@ -24,21 +26,22 @@ public class ProfessionalController {
 
     @RequestMapping("get-by-id/{id}")
     @ResponseBody
-    public ProfessionalDTO getById(@PathVariable("id")Long id){
+    public ResponseEntity<ProfessionalDTO> getById(@PathVariable("id")Long id){
         ProfessionalDTO pDTO = null;
         try{
             Professional p = this.proDAO.getById(id);
             pDTO = MapStructMapper.INSTANCE.professionalToProfessionalDTO(p);
         }catch (Exception ex) {
             System.out.println(ex);
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
-        return pDTO;
+        return new ResponseEntity<>(pDTO,HttpStatus.FOUND);
     }
 
 
     @RequestMapping("get-by-email/{email}")
     @ResponseBody
-    public ProfessionalDTO getByEmail(@PathVariable("email") String mail) {
+    public ResponseEntity<ProfessionalDTO> getByEmail(@PathVariable("email") String mail) {
         ProfessionalDTO pDTO = null;
         try {
             Professional p = this.proDAO.getByMail(mail);
@@ -46,36 +49,52 @@ public class ProfessionalController {
         }
         catch (Exception ex) {
             System.out.println(ex);
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
-        return pDTO;
+        return new ResponseEntity<>(pDTO,HttpStatus.FOUND);
     }
 
     @RequestMapping("get-all-professional")
     @ResponseBody
-    public List<ProfessionalDTO> getAllProfessional(){
+    public ResponseEntity<List<ProfessionalDTO>> getAllProfessional(){
         List<ProfessionalDTO> pDTO = new ArrayList<>();
         try {
             for(Professional p : this.proDAO.findAll()){
                 pDTO.add(MapStructMapper.INSTANCE.professionalToProfessionalDTO(p));
             }
-            return pDTO;
+            return new ResponseEntity<>(pDTO,HttpStatus.FOUND);
         } catch (Exception ex) {
             System.out.println(ex);
-            return null;
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PostMapping("create")
-    public Professional createProfessional(@RequestBody Professional pro){
-        proDAO.save(pro);
-        return pro;
+    public ResponseEntity<ProfessionalDTO> createProfessional(@RequestBody ProfessionalDTO proDTO){
+        try{
+            Professional pro = MapStructMapper.INSTANCE.professionalDTOToProfessional(proDTO);
+            proDAO.saveAndFlush(pro);
+        }catch (Exception ex){
+            Logger.getGlobal().warning(ex.toString());
+            return new ResponseEntity<>(proDTO,HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        return new ResponseEntity<>(proDTO,HttpStatus.CREATED);
     }
 
     @DeleteMapping("delete/{id}")
-    public String deleteProfessional(@PathVariable("id") long id){
-        proDAO.deleteById(id);
-        return "proffessional deleted";
+    public ResponseEntity<String> deleteProfessional(@PathVariable("id") long id){
+        try{
+            if(proDAO.getById(id) != null){
+                proDAO.deleteById(id);
+                return new ResponseEntity<>("Patiend deleted", HttpStatus.ACCEPTED);
+            }else{
+                return new ResponseEntity<>("Professional does not exist", HttpStatus.NOT_FOUND);
+            }
+        }catch (Exception e){
+            Logger.getGlobal().warning(e.toString());
+            return new ResponseEntity<>("Error", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
-
-
 }
